requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-cors-headers==3.13.0
django-import-export==2.8.0
djangorestframework==3.13.1
drf-yasg==1.21.3
et-xmlfile==1.1.0
google-api-core==2.8.2
google-auth==2.10.0
googleapis-common-protos==1.56.4
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zope.interface==5.4.0
